+++ nf_core/modules/lint/__init__.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 """
 Code for linting modules in the nf-core/modules repository and
 in nf-core pipelines
@@ -72,7 +71,14 @@ class ModuleLint(ComponentCommand):
         no_pull=False,
         hide_progress=False,
     ):
-        super().__init__("modules", dir=dir, remote_url=remote_url, branch=branch, no_pull=no_pull, hide_progress=False)
+        super().__init__(
+            "modules",
+            dir=dir,
+            remote_url=remote_url,
+            branch=branch,
+            no_pull=no_pull,
+            hide_progress=hide_progress,
+        )
 
         self.fail_warned = fail_warned
         self.passed = []
@@ -133,7 +139,6 @@ class ModuleLint(ComponentCommand):
         module=None,
         key=(),
         all_modules=False,
-        hide_progress=False,
         print_results=True,
         show_passed=False,
         local=False,