+++ tests/modules/update.py
@@ -2,13 +2,17 @@ import filecmp
 import os
 import shutil
 import tempfile
+from pathlib import Path
+from unittest import mock
 
+import questionary
 import yaml
 
 import nf_core.utils
 from nf_core.modules.install import ModuleInstall
 from nf_core.modules.modules_json import ModulesJson
 from nf_core.modules.modules_repo import NF_CORE_MODULES_NAME, NF_CORE_MODULES_REMOTE
+from nf_core.modules.patch import ModulePatch
 from nf_core.modules.update import ModuleUpdate
 
 from ..utils import (
@@ -41,7 +45,9 @@ def test_install_and_update(self):
 def test_install_at_hash_and_update(self):
     """Installs an old version of a module in the pipeline and updates it"""
     assert self.mods_install_old.install("trimgalore")
-    update_obj = ModuleUpdate(self.pipeline_dir, show_diff=False, remote_url=GITLAB_URL, branch=OLD_TRIMGALORE_BRANCH)
+    update_obj = ModuleUpdate(
+        self.pipeline_dir, show_diff=False, update_deps=True, remote_url=GITLAB_URL, branch=OLD_TRIMGALORE_BRANCH
+    )
 
     # Copy the module files and check that they are affected by the update
     tmpdir = tempfile.mkdtemp()
@@ -225,7 +231,7 @@ def test_update_different_branch_single_module(self):
     assert install_obj.install("fastp")
 
     update_obj = ModuleUpdate(
-        self.pipeline_dir, remote_url=GITLAB_URL, branch=GITLAB_BRANCH_TEST_BRANCH, show_diff=False
+        self.pipeline_dir, update_deps=True, remote_url=GITLAB_URL, branch=GITLAB_BRANCH_TEST_BRANCH, show_diff=False
     )
     update_obj.update("fastp")
 
@@ -286,6 +292,72 @@ def test_update_different_branch_mix_modules_branch_test(self):
         == GITLAB_BRANCH_TEST_BRANCH
     )
     assert modules_json.get_module_version("multiqc", GITLAB_URL, GITLAB_REPO) == GITLAB_BRANCH_TEST_NEW_SHA
+
+
+# Mock questionary answer: do not update module, only show diffs
+@mock.patch.object(questionary.Question, "unsafe_ask", return_value=False)
+def test_update_only_show_differences(self, mock_prompt):
+    """Try updating all modules showing differences.
+    Don't update some of them.
+    Check that the sha in modules.json is not changed."""
+    modules_json = ModulesJson(self.pipeline_dir)
+    update_obj = ModuleUpdate(self.pipeline_dir, update_all=True, show_diff=True)
+
+    tmpdir = tempfile.mkdtemp()
+    shutil.rmtree(tmpdir)
+    shutil.copytree(Path(self.pipeline_dir, "modules", NF_CORE_MODULES_NAME), tmpdir)
+
+    assert update_obj.update() is True
+
+    mod_json = modules_json.get_modules_json()
+    # Loop through all modules and check that they are NOT updated (according to the modules.json file)
+    # Modules that can be updated but shouldn't are custom/dumpsoftwareversions and fastqc
+    # Module multiqc is already up to date so don't check
+    for mod in ["custom/dumpsoftwareversions", "fastqc"]:
+        correct_git_sha = list(update_obj.modules_repo.get_component_git_log(mod, "modules", depth=1))[0]["git_sha"]
+        current_git_sha = mod_json["repos"][NF_CORE_MODULES_REMOTE]["modules"][NF_CORE_MODULES_NAME][mod]["git_sha"]
+        assert correct_git_sha != current_git_sha
+        assert cmp_module(Path(tmpdir, mod), Path(self.pipeline_dir, "modules", NF_CORE_MODULES_NAME, mod)) is True
+
+
+# Mock questionary answer: do not update module, only show diffs
+@mock.patch.object(questionary.Question, "unsafe_ask", return_value=False)
+def test_update_only_show_differences_when_patch(self, mock_prompt):
+    """Try updating all modules showing differences when there's a patched module.
+    Don't update some of them.
+    Check that the sha in modules.json is not changed."""
+    modules_json = ModulesJson(self.pipeline_dir)
+    update_obj = ModuleUpdate(self.pipeline_dir, update_all=True, show_diff=True)
+
+    # Modify fastqc module, it will have a patch which will be applied during update
+    # We modify fastqc because it's one of the modules that can be updated and there's another one before it (custom/dumpsoftwareversions)
+    module_path = Path(self.pipeline_dir, "modules", "nf-core", "fastqc")
+    main_path = Path(module_path, "main.nf")
+    with open(main_path, "r") as fh:
+        lines = fh.readlines()
+    for line_index in range(len(lines)):
+        if lines[line_index] == "    label 'process_medium'\n":
+            lines[line_index] = "    label 'process_low'\n"
+            break
+    with open(main_path, "w") as fh:
+        fh.writelines(lines)
+    # Create a patch file
+    patch_obj = ModulePatch(self.pipeline_dir)
+    patch_obj.patch("fastqc")
+    # Check that a patch file with the correct name has been created
+    assert set(os.listdir(module_path)) == {"main.nf", "meta.yml", "fastqc.diff"}
+
+    # Update all modules
+    assert update_obj.update() is True
+
+    mod_json = modules_json.get_modules_json()
+    # Loop through all modules and check that they are NOT updated (according to the modules.json file)
+    # Modules that can be updated but shouldn't are custom/dumpsoftwareversions and fastqc
+    # Module multiqc is already up to date so don't check
+    for mod in ["custom/dumpsoftwareversions", "fastqc"]:
+        correct_git_sha = list(update_obj.modules_repo.get_component_git_log(mod, "modules", depth=1))[0]["git_sha"]
+        current_git_sha = mod_json["repos"][NF_CORE_MODULES_REMOTE]["modules"][NF_CORE_MODULES_NAME][mod]["git_sha"]
+        assert correct_git_sha != current_git_sha
 
 
 def cmp_module(dir1, dir2):