+++ tests/utils.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
 """
 Helper functions for tests
 """
@@ -16,6 +14,7 @@ GITLAB_URL = "https://gitlab.com/nf-core/modules-test.git"
 GITLAB_REPO = "nf-core"
 GITLAB_DEFAULT_BRANCH = "main-restructure"
 GITLAB_SUBWORKFLOWS_BRANCH = "subworkflows"
+OLD_SUBWORKFLOWS_SHA = "f3c078809a2513f1c95de14f6633fe1f03572fdb"
 # Branch test stuff
 GITLAB_BRANCH_TEST_BRANCH = "branch-tester-restructure"
 GITLAB_BRANCH_TEST_OLD_SHA = "bce3f17980b8d1beae5e917cfd3c65c0c69e04b5"